Return Future<void> rather than void for pushContext
To circumvent potential race conditions with calling getContext too soon
after pushContext, we'd like to return a Future from pushContext and
also ensure that the key is set therein before returning.

Alternatively, we could just block the thread until it returns properly
... not sure if that would be a good idea though, maybe a Future is
safer?

Still todo is to figure out the basics as to when to mark a Future done
or failed, there are ways to do this in other languages but not sure
about Java.